refactor(bayn): decompose observe mutation orchestration
