+++ services/bayn/src/observe-composition/mutation-decisions.ts
@@ -1,0 +1,402 @@
+import { Result } from 'effect'
+
+import { MutationOperation } from '../broker/alpaca-mutations'
+import { CycleTerminalReason } from '../cycle'
+import {
+  Authority,
+  IntentState,
+  OrderStatus,
+  ReconciliationStatus,
+  TerminalOutcome,
+  type Intent,
+  type Order,
+  type Position,
+} from '../execution/contracts'
+import { MutationEventType, type MutationEvent } from '../execution/mutations'
+import type { PlannedTargetQuantity } from '../target-planner'
+
+const quantityScale = 1_000_000n
+
+const comparePositionSymbol = (left: Position, right: Position): number =>
+  left.symbol < right.symbol ? -1 : left.symbol > right.symbol ? 1 : 0
+
+const divideAwayFromZero = (numerator: bigint): bigint => {
+  const magnitude = numerator < 0n ? -numerator : numerator
+  const rounded = magnitude === 0n ? 0n : (magnitude + quantityScale - 1n) / quantityScale
+  return numerator < 0n ? -rounded : rounded
+}
+
+const projectMutationPosition = (
+  positions: readonly Position[],
+  target: PlannedTargetQuantity,
+  accountId: string,
+  observedAt: string,
+): readonly Position[] => {
+  const retained = positions.filter((position) => position.symbol !== target.symbol)
+  const quantity = BigInt(target.targetQuantityMicros)
+  if (quantity === 0n) return retained
+  const previous = positions.find((position) => position.symbol === target.symbol)
+  const referencePrice = BigInt(target.referencePriceMicros)
+  const averageEntryPrice = BigInt(previous?.averageEntryPriceMicros ?? target.referencePriceMicros)
+  const projected: Position = {
+    schemaVersion: 'bayn.paper-position.v1',
+    accountId: previous?.accountId ?? accountId,
+    symbol: target.symbol,
+    quantityMicros: target.targetQuantityMicros,
+    averageEntryPriceMicros: averageEntryPrice.toString(),
+    marketPriceMicros: target.referencePriceMicros,
+    marketValueMicros: divideAwayFromZero(quantity * referencePrice).toString(),
+    unrealizedPnlMicros: divideAwayFromZero(quantity * (referencePrice - averageEntryPrice)).toString(),
+    observedAt: previous?.observedAt ?? observedAt,
+  }
+  return [...retained, projected].sort(comparePositionSymbol)
+}
+
+export const projectWorstCasePendingMutationPosition = (
+  positions: readonly Position[],
+  target: PlannedTargetQuantity,
+  accountId: string,
+  observedAt: string,
+): readonly Position[] => {
+  const current = positions.find((position) => position.symbol === target.symbol)
+  const currentQuantity = current === undefined ? 0n : BigInt(current.quantityMicros)
+  const targetQuantity = BigInt(target.targetQuantityMicros)
+  return targetQuantity <= currentQuantity
+    ? positions
+    : projectMutationPosition(positions, target, accountId, observedAt)
+}
+
+export type PreparedMutationIntentDecision =
+  | { readonly _tag: 'Submit' }
+  | { readonly _tag: 'Recover'; readonly eventType: MutationEventType }
+  | { readonly _tag: 'Pending'; readonly order: Order }
+  | { readonly _tag: 'SkipTerminal' }
+
+export type PreparedMutationIntentDecisionFailure = {
+  readonly _tag: 'PreparedMutationIntentDecisionFailure'
+  readonly intentId: string
+  readonly message: string
+  readonly eventType?: MutationEventType
+}
+
+export const decidePreparedMutationIntent = (
+  intent: Intent,
+  latest: MutationEvent | undefined,
+): Result.Result<PreparedMutationIntentDecision, PreparedMutationIntentDecisionFailure> => {
+  if (intent.state === IntentState.Terminal) return Result.succeed({ _tag: 'SkipTerminal' })
+  if (latest === undefined) return Result.succeed({ _tag: 'Submit' })
+  switch (latest.eventType) {
+    case MutationEventType.SubmitAccepted:
+    case MutationEventType.RecoveryFound:
+      return latest.brokerOrderId === undefined
+        ? Result.fail({
+            _tag: 'PreparedMutationIntentDecisionFailure',
+            intentId: intent.intentId,
+            eventType: latest.eventType,
+            message: 'accepted nonterminal submit lacks a durable broker order identity',
+          })
+        : Result.succeed({
+            _tag: 'Pending',
+            order: {
+              schemaVersion: 'bayn.paper-order.v1',
+              accountId: intent.accountId,
+              brokerOrderId: latest.brokerOrderId,
+              clientOrderId: intent.clientOrderId,
+              intentId: intent.intentId,
+              symbol: intent.symbol,
+              side: intent.side,
+              orderType: intent.orderType,
+              timeInForce: intent.timeInForce,
+              quantityMicros: intent.quantityMicros,
+              filledQuantityMicros: '0',
+              status: OrderStatus.New,
+              observedAt: latest.occurredAt,
+            },
+          })
+    case MutationEventType.SubmitRejected:
+    case MutationEventType.SubmitDenied:
+      return Result.fail({
+        _tag: 'PreparedMutationIntentDecisionFailure',
+        intentId: intent.intentId,
+        eventType: latest.eventType,
+        message: 'terminal submit event does not match the nonterminal durable intent state',
+      })
+    case MutationEventType.SubmitStarted:
+    case MutationEventType.SubmitUnknown:
+    case MutationEventType.RecoveryNotFound:
+    case MutationEventType.RecoveryUnknown:
+    case MutationEventType.CancelStarted:
+    case MutationEventType.CancelAccepted:
+    case MutationEventType.CancelUnknown:
+      return Result.succeed({ _tag: 'Recover', eventType: latest.eventType })
+  }
+}
+
+export type PreparedMutationIntentAdmissionFailure = {
+  readonly _tag: 'PreparedMutationIntentAdmissionFailure'
+  readonly reason:
+    | 'accounting-inexact'
+    | 'authority'
+    | 'expiry'
+    | 'reconciliation-not-exact'
+    | 'unknown-mutation'
+    | 'unknown-order'
+  readonly message: string
+}
+
+export const decidePreparedMutationIntentAdmission = (
+  prepared: PreparedMutationIntentDecision,
+  effectiveAuthority: Authority,
+  observedAt: string,
+  expiresAt: string,
+  unknownMutationCount: number,
+  reconciliationStatus: ReconciliationStatus = ReconciliationStatus.Exact,
+  accountingExact = true,
+  unknownOrderCount = 0,
+): Result.Result<void, PreparedMutationIntentAdmissionFailure> => {
+  if (prepared._tag !== 'Submit') return Result.succeed(undefined)
+  if (effectiveAuthority !== Authority.Paper) {
+    return Result.fail({
+      _tag: 'PreparedMutationIntentAdmissionFailure',
+      reason: 'authority',
+      message: 'fresh PAPER submit requires current effective PAPER authority',
+    })
+  }
+  if (observedAt >= expiresAt) {
+    return Result.fail({
+      _tag: 'PreparedMutationIntentAdmissionFailure',
+      reason: 'expiry',
+      message: 'fresh PAPER submit is forbidden at or after decision expiry',
+    })
+  }
+  if (reconciliationStatus !== ReconciliationStatus.Exact) {
+    return Result.fail({
+      _tag: 'PreparedMutationIntentAdmissionFailure',
+      reason: 'reconciliation-not-exact',
+      message: 'fresh PAPER submit requires exact same-pass reconciliation',
+    })
+  }
+  if (!accountingExact) {
+    return Result.fail({
+      _tag: 'PreparedMutationIntentAdmissionFailure',
+      reason: 'accounting-inexact',
+      message: 'fresh PAPER submit requires exact same-pass accounting',
+    })
+  }
+  if (unknownMutationCount !== 0) {
+    return Result.fail({
+      _tag: 'PreparedMutationIntentAdmissionFailure',
+      reason: 'unknown-mutation',
+      message: 'fresh PAPER submit is forbidden while a mutation outcome is unknown',
+    })
+  }
+  if (unknownOrderCount !== 0) {
+    return Result.fail({
+      _tag: 'PreparedMutationIntentAdmissionFailure',
+      reason: 'unknown-order',
+      message: 'fresh PAPER submit is forbidden while a broker order is unknown',
+    })
+  }
+  return Result.succeed(undefined)
+}
+
+export const appendPendingMutationOrder = (orders: readonly Order[], pending: Order): readonly Order[] =>
+  orders.some((order) => order.brokerOrderId === pending.brokerOrderId || order.clientOrderId === pending.clientOrderId)
+    ? orders
+    : [...orders, pending]
+
+export interface PaperCycleIntentTerminalEvidence {
+  readonly state: IntentState
+  readonly terminalOutcome?: TerminalOutcome
+  readonly updatedAt: string
+  readonly latestMutationAt?: string
+}
+
+export interface PaperCycleReconciliationEvidence {
+  readonly status: ReconciliationStatus
+  readonly reconciledAt: string
+  readonly accountingExact: boolean
+  readonly unknownMutationCount: number
+  readonly unknownOrderCount: number
+}
+
+export type PaperCycleCompletionDecision =
+  | { readonly _tag: 'Complete' }
+  | {
+      readonly _tag: 'Wait'
+      readonly reason:
+        | 'accounting-inexact'
+        | 'intent-nonterminal'
+        | 'intent-unsuccessful'
+        | 'reconciliation-not-later'
+        | 'reconciliation-not-exact'
+        | 'unknown-mutation'
+        | 'unknown-order'
+    }
+
+export const decidePaperCycleCompletion = (
+  documentCreatedAt: string,
+  intents: readonly PaperCycleIntentTerminalEvidence[],
+  reconciliation: PaperCycleReconciliationEvidence,
+): PaperCycleCompletionDecision => {
+  if (intents.some((intent) => intent.state !== IntentState.Terminal)) {
+    return { _tag: 'Wait', reason: 'intent-nonterminal' }
+  }
+  if (intents.some((intent) => intent.terminalOutcome !== TerminalOutcome.Filled)) {
+    return { _tag: 'Wait', reason: 'intent-unsuccessful' }
+  }
+  if (reconciliation.status !== ReconciliationStatus.Exact) {
+    return { _tag: 'Wait', reason: 'reconciliation-not-exact' }
+  }
+  if (!reconciliation.accountingExact) return { _tag: 'Wait', reason: 'accounting-inexact' }
+  if (reconciliation.unknownMutationCount !== 0) return { _tag: 'Wait', reason: 'unknown-mutation' }
+  if (reconciliation.unknownOrderCount !== 0) return { _tag: 'Wait', reason: 'unknown-order' }
+  const latestEvidenceAt = Math.max(
+    Date.parse(documentCreatedAt),
+    ...intents.flatMap((intent) => [
+      Date.parse(intent.updatedAt),
+      Date.parse(intent.latestMutationAt ?? intent.updatedAt),
+    ]),
+  )
+  if (Date.parse(reconciliation.reconciledAt) <= latestEvidenceAt) {
+    return { _tag: 'Wait', reason: 'reconciliation-not-later' }
+  }
+  return { _tag: 'Complete' }
+}
+
+export type PreparedMutationRecoveryDecision =
+  | { readonly _tag: 'NoRecovery' }
+  | {
+      readonly _tag: 'Recover'
+      readonly operation: MutationOperation
+      readonly event: MutationEvent
+    }
+
+export const decidePreparedMutationRecovery = (
+  intent: Intent,
+  latestSubmit: MutationEvent | undefined,
+  latestCancel: MutationEvent | undefined,
+): Result.Result<PreparedMutationRecoveryDecision, PreparedMutationIntentDecisionFailure> => {
+  if (latestCancel !== undefined) {
+    if (latestCancel.operation !== MutationOperation.Cancel) {
+      return Result.fail({
+        _tag: 'PreparedMutationIntentDecisionFailure',
+        intentId: intent.intentId,
+        eventType: latestCancel.eventType,
+        message: 'durable cancellation recovery read returned a non-cancel mutation',
+      })
+    }
+    return intent.state === IntentState.Terminal && latestCancel.eventType === MutationEventType.RecoveryFound
+      ? Result.succeed({ _tag: 'NoRecovery' })
+      : Result.succeed({ _tag: 'Recover', operation: MutationOperation.Cancel, event: latestCancel })
+  }
+  if (latestSubmit === undefined) return Result.succeed({ _tag: 'NoRecovery' })
+  if (latestSubmit.operation !== MutationOperation.Submit) {
+    return Result.fail({
+      _tag: 'PreparedMutationIntentDecisionFailure',
+      intentId: intent.intentId,
+      eventType: latestSubmit.eventType,
+      message: 'durable submit recovery read returned a non-submit mutation',
+    })
+  }
+  if (
+    intent.state === IntentState.Terminal &&
+    (latestSubmit.eventType === MutationEventType.SubmitAccepted ||
+      latestSubmit.eventType === MutationEventType.SubmitRejected ||
+      latestSubmit.eventType === MutationEventType.SubmitDenied ||
+      latestSubmit.eventType === MutationEventType.RecoveryFound)
+  ) {
+    return Result.succeed({ _tag: 'NoRecovery' })
+  }
+  if (
+    intent.state !== IntentState.Terminal &&
+    (latestSubmit.eventType === MutationEventType.SubmitRejected ||
+      latestSubmit.eventType === MutationEventType.SubmitDenied)
+  ) {
+    return Result.fail({
+      _tag: 'PreparedMutationIntentDecisionFailure',
+      intentId: intent.intentId,
+      eventType: latestSubmit.eventType,
+      message: 'terminal submit event does not match the nonterminal durable intent state',
+    })
+  }
+  return Result.succeed({ _tag: 'Recover', operation: MutationOperation.Submit, event: latestSubmit })
+}
+
+export type PreparedMutationCycleStep =
+  | { readonly _tag: 'RunCycle' }
+  | {
+      readonly _tag: 'Execute'
+      readonly action: 'RECOVER_SUBMIT' | 'RECOVER_CANCEL'
+      readonly intentId: string
+      readonly observedAt: string
+    }
+  | {
+      readonly _tag: 'Execute'
+      readonly action: 'SUBMIT'
+      readonly intentId: string
+      readonly observedAt: string
+      readonly submitExpiresAt: string
+    }
+  | {
+      readonly _tag: 'Block'
+      readonly reason:
+        | CycleTerminalReason.MissedSubmission
+        | CycleTerminalReason.ProvenanceMismatch
+        | CycleTerminalReason.Risk
+      readonly observedAt: string
+    }
+  | { readonly _tag: 'Wait'; readonly observedAt: string }
+  | { readonly _tag: 'Complete'; readonly observedAt: string }
+
+export type BoundMutationCycleOutcome = Exclude<PreparedMutationCycleStep, { readonly _tag: 'Execute' }>
+
+export const mutationRecoveryIsDue = (event: MutationEvent, observedAt: string): boolean =>
+  Date.parse(observedAt) >= Date.parse(event.occurredAt) + event.consistencyDelayMs
+
+export const paperSubmitExpiresAt = (documentExpiresAt: string, riskDecisionExpiresAt: string): string =>
+  riskDecisionExpiresAt < documentExpiresAt ? riskDecisionExpiresAt : documentExpiresAt
+
+export const expiredPaperPlanTerminalReason = (
+  observedAt: string,
+  submitExpiresAt: string,
+  submissionCutoffAt: string,
+): CycleTerminalReason.MissedSubmission | CycleTerminalReason.Risk | undefined =>
+  observedAt < submitExpiresAt
+    ? undefined
+    : observedAt >= submissionCutoffAt
+      ? CycleTerminalReason.MissedSubmission
+      : CycleTerminalReason.Risk
+
+export type MutationIntentSettlementDecision =
+  | {
+      readonly _tag: 'Settled'
+      readonly outcome: 'accepted' | 'rejected' | 'denied'
+    }
+  | {
+      readonly _tag: 'Unresolved'
+      readonly eventType: MutationEventType
+    }
+
+export const decideMutationIntentSettlement = (eventType: MutationEventType): MutationIntentSettlementDecision => {
+  switch (eventType) {
+    case MutationEventType.SubmitAccepted:
+    case MutationEventType.RecoveryFound:
+      return { _tag: 'Settled', outcome: 'accepted' }
+    case MutationEventType.SubmitRejected:
+      return { _tag: 'Settled', outcome: 'rejected' }
+    case MutationEventType.SubmitDenied:
+      return { _tag: 'Settled', outcome: 'denied' }
+    default:
+      return { _tag: 'Unresolved', eventType }
+  }
+}
+
+export interface MutationIntentExecutionResult {
+  readonly settlement: Extract<MutationIntentSettlementDecision, { readonly _tag: 'Settled' }>
+  readonly consistencyDelayMs: number
+  readonly operation: MutationOperation
+}
+
+export const mutationIntentReconciliationDelayMs = (result: MutationIntentExecutionResult): number =>
+  result.settlement.outcome === 'accepted' ? result.consistencyDelayMs : 0
